refactor : Order Entity 수정, domain -> model 폴더 명 변경
Order Entity에 deletedAt (소프트 삭제 용) 추가
