--- order-service/src/main/java/com/example/order_service/domain/OrderItemOption.java
@@ -1,20 +1,0 @@
-package com.example.order_service.domain;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@Getter
-public class OrderItemOption {
-
-    private Long optionId;
-
-    private String optionName;
-
-    private int price;
-
-}
