--- order-service/src/main/java/com/example/order_service/domain/OrderStatus.java
@@ -1,5 +1,0 @@
-package com.example.order_service.domain;
-
-public enum OrderStatus {
-    ORDERED, COOKING, DELIVERING, COMPLETED, CANCELED
-}
