+++ order-service/src/main/java/com/example/order_service/model/OrderItemOption.java
@@ -1,4 +1,4 @@
-package com.example.order_service.domain;
+package com.example.order_service.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
